requirements_backup.txt
addict==2.4.0
attrs==23.2.0
backcall==0.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.4
ConfigArgParse==1.7
cycler==0.11.0
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
-e git+https://github.com/ingra14m/depth-diff-gaussian-rasterization@e49506654e8e11ed8a62d22bcb693e943fdecacf#egg=diff_gaussian_rasterization&subdirectory=../../../submodules/depth-diff-gaussian-rasterization
fastjsonschema==2.19.1
Flask==2.2.5
fonttools==4.38.0
idna==3.7
imageio==2.31.2
imageio-ffmpeg==0.4.9
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipython==7.34.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.17.3
jupyter_core==4.12.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lpips==0.1.4
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mmcv==1.6.0
nbformat==5.7.0
nest-asyncio==1.6.0
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
open3d==0.17.0
opencv-python==4.9.0.80
packaging==24.0
pandas==1.3.5
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
plotly==5.18.0
plyfile==0.9
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
Pygments==2.17.2
pyparsing==3.1.2
pyquaternion==0.9.9
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
pytz==2024.1
PyWavelets==1.3.0
PyYAML==6.0.1
requests==2.31.0
retrying==1.3.4
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
-e git+https://gitlab.inria.fr/bkerbl/simple-knn.git@44f764299fa305faf6ec5ebd99939e0508331503#egg=simple_knn&subdirectory=../../../submodules/simple-knn
six==1.16.0
tenacity==8.2.3
threadpoolctl==3.1.0
tifffile==2021.11.2
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.66.4
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
Werkzeug==2.2.3
widgetsnbextension==4.0.10
yapf==0.40.2
zipp==3.15.0
